+++ src/Dependencies.lua
@@ -23,10 +23,10 @@ require('src.constants')
 
 
 require('src.states.BaseState')
-require('src.states.entity.CustomerState')
 require('src.states.game.PlayState')
 require('src.states.game.StartMenu')
 require('src.states.game.DayEndState')
+require('src.states.game.PauseMenu')
 
 require('src.states.entity.CoffeeMachine')
 require('src.states.entity.CustomerState')